Update search criteria for excluding Defender signature updates to sue new name
The search criteria used to exclude Defender signature updates from the update date check is using an old name. Microsoft changed the name of the update from Definition Update to Security Intelligence Update in 2019. As a result, the script is not excluding these updates and the date shown doesn't reflect the last non-Defender update.

This fix matches on both terms to cover the situation where an older system with the old term is being checked.